refactor(frontend): reduce complexity of state management in calculateur indemnite
